Modify to move some functions to utils.js
Modify to move some functions to utils.js

Modify to move some functions to utils.js

Modify to move some functions to utils.js

Modify to move some functions to utils.js

Modify to move some functions to utils.js

Modify to move some functions to utils.js

Modify to move some functions to utils.js
